lib: utils/irqchip: Use IO pointer return from ioreamp()
addr in struct aplic_data,  struct imsic_regs and struct plic_data are now a IO
pointer return from ioremap() to make irqchip driver compatible with
Zcheripurecap ABI.

+++ lib/utils/irqchip/aplic.c
@@ -289,7 +289,7 @@ int aplic_cold_irqchip_init(struct aplic_data *aplic)
 	    ((first_deleg_irq < last_deleg_irq) &&
 	    (last_deleg_irq == aplic->num_source) &&
 	    (first_deleg_irq == 1))) {
-		rc = sbi_domain_root_add_memrange(aplic->addr, aplic->size, PAGE_SIZE,
+		rc = sbi_domain_root_add_memrange((unsigned long)aplic->addr, aplic->size, PAGE_SIZE,
 						  SBI_DOMAIN_MEMREGION_MMIO |
 						  SBI_DOMAIN_MEMREGION_M_READABLE |
 						  SBI_DOMAIN_MEMREGION_M_WRITABLE);
